Fixed bug in warn(), renamed longenum to enum_long

+++ Emulator/VICII/VICII.cpp
@@ -445,12 +445,12 @@ VICII::_inspect()
 void
 VICII::_dumpConfig()
 {
-    msg("    Chip model : %ld (%s)\n", (long)config.revision, vicRevisionName(config.revision));
+    msg("    Chip model : %lld (%s)\n", config.revision, vicRevisionName(config.revision));
     msg("  Gray dot bug : %s\n", config.grayDotBug ? "yes" : "no");
     msg("           PAL : %s\n", isPAL() ? "yes" : "no");
     msg("          NTSC : %s\n", isNTSC() ? "yes" : "no");
     msg("is656x, is856x : %d %d\n", is656x(), is856x());
-    msg("    Glue logic : %ld (%s)\n", (long)config.glueLogic, glueLogicName(config.glueLogic));
+    msg("    Glue logic : %lld (%s)\n", config.glueLogic, glueLogicName(config.glueLogic));
 }
 
 void
